data/requirements.txt
beautifulsoup4==4.12.3
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.12.0
elasticsearch==8.12.1
idna==3.6
isort==5.13.2
joblib==1.3.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.1.0
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
py4j==0.10.9.7
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
